requirements.txt
Django==4.2
djangorestframework==3.14.0
django-filter==23.2
pytest-django==4.5.2
pytest==7.4.0
black==23.7.0
gunicorn==20.1.0
whitenoise==6.5.0
